Created temp PNF/Datadog specific log display.
